Share Apple Pay component delegate handler
