+++ ios/adyen_checkout/Sources/adyen_checkout/components/applePay/ApplePayAdvancedComponent.swift
@@ -3,13 +3,13 @@
     import AdyenComponents
 #endif
 import Foundation
-import PassKit
 
 class ApplePayAdvancedComponent: BaseApplePayComponent {
     private let componentFlutterApi: ComponentFlutterInterface
     private let configuration: InstantPaymentConfigurationDTO
     private let paymentMethodResponse: String
     private let componentId: String
+    private var applePayComponentDelegateHandler: ApplePayComponentDelegateHandler?
 
     init(
         componentFlutterApi: ComponentFlutterInterface,
@@ -60,7 +60,12 @@ class ApplePayAdvancedComponent: BaseApplePayComponent {
         let applePayComponent = try ApplePayComponent(paymentMethod: paymentMethod, context: context, configuration: configuration)
         applePayComponent.delegate = self
         if self.configuration.applePayConfigurationDTO?.hasOnShippingMethodChange == true {
-            applePayComponent.applePayDelegate = self
+            let applePayComponentDelegateHandler = ApplePayComponentDelegateHandler(
+                componentFlutterApi: componentFlutterApi,
+                componentId: componentId
+            )
+            self.applePayComponentDelegateHandler = applePayComponentDelegateHandler
+            applePayComponent.applePayDelegate = applePayComponentDelegateHandler
         }
         return applePayComponent
     }
@@ -149,44 +154,5 @@ extension ApplePayAdvancedComponent: PaymentComponentDelegate {
             componentCommunicationModel: componentCommunicationModel,
             completion: { _ in }
         )
-    }
-}
-
-extension ApplePayAdvancedComponent: ApplePayComponentDelegate {
-    func didUpdate(
-        contact: PKContact,
-        for payment: ApplePayPayment,
-        completion: @escaping (PKPaymentRequestShippingContactUpdate) -> Void
-    ) {
-        completion(PKPaymentRequestShippingContactUpdate(paymentSummaryItems: payment.summaryItems))
-    }
-
-    func didUpdate(
-        shippingMethod: PKShippingMethod,
-        for payment: ApplePayPayment,
-        completion: @escaping (PKPaymentRequestShippingMethodUpdate) -> Void
-    ) {
-        componentFlutterApi.onApplePayShippingMethodChange(
-            componentId: componentId,
-            shippingMethod: shippingMethod.toDTO(currencyCode: payment.currencyCode),
-            currentSummaryItems: payment.summaryItems.map { $0.toDTO(currencyCode: payment.currencyCode) },
-            completion: { result in
-                switch result {
-                case let .success(update):
-                    completion(update.toPKPaymentRequestShippingMethodUpdate())
-                case .failure:
-                    completion(PKPaymentRequestShippingMethodUpdate(paymentSummaryItems: payment.summaryItems))
-                }
-            }
-        )
-    }
-
-    @available(iOS 15.0, *)
-    func didUpdate(
-        couponCode: String,
-        for payment: ApplePayPayment,
-        completion: @escaping (PKPaymentRequestCouponCodeUpdate) -> Void
-    ) {
-        completion(PKPaymentRequestCouponCodeUpdate(paymentSummaryItems: payment.summaryItems))
     }
 }
