+++ ios/adyen_checkout/Sources/adyen_checkout/components/applePay/ApplePaySessionComponent.swift
@@ -2,13 +2,13 @@
 #if canImport(AdyenComponents)
     import AdyenComponents
 #endif
-import PassKit
 
 class ApplePaySessionComponent: BaseApplePayComponent {
     private let sessionHolder: SessionHolder
     private let configuration: InstantPaymentConfigurationDTO
     private let componentFlutterApi: ComponentFlutterInterface
     private let componentId: String
+    private var applePayComponentDelegateHandler: ApplePayComponentDelegateHandler?
     
     init(
         sessionHolder: SessionHolder,
@@ -44,7 +44,12 @@ class ApplePaySessionComponent: BaseApplePayComponent {
         let applePayComponent = try ApplePayComponent(paymentMethod: paymentMethod, context: context, configuration: configuration)
         applePayComponent.delegate = sessionHolder.session
         if self.configuration.applePayConfigurationDTO?.hasOnShippingMethodChange == true {
-            applePayComponent.applePayDelegate = self
+            let applePayComponentDelegateHandler = ApplePayComponentDelegateHandler(
+                componentFlutterApi: componentFlutterApi,
+                componentId: componentId
+            )
+            self.applePayComponentDelegateHandler = applePayComponentDelegateHandler
+            applePayComponent.applePayDelegate = applePayComponentDelegateHandler
         }
         setupSessionFlowDelegate()
         return applePayComponent
@@ -65,44 +70,5 @@ class ApplePaySessionComponent: BaseApplePayComponent {
         super.finalizeAndDismissComponent(success: success, completion: { [weak self] in
             completion()
         })
-    }
-}
-
-extension ApplePaySessionComponent: ApplePayComponentDelegate {
-    func didUpdate(
-        contact: PKContact,
-        for payment: ApplePayPayment,
-        completion: @escaping (PKPaymentRequestShippingContactUpdate) -> Void
-    ) {
-        completion(PKPaymentRequestShippingContactUpdate(paymentSummaryItems: payment.summaryItems))
-    }
-
-    func didUpdate(
-        shippingMethod: PKShippingMethod,
-        for payment: ApplePayPayment,
-        completion: @escaping (PKPaymentRequestShippingMethodUpdate) -> Void
-    ) {
-        componentFlutterApi.onApplePayShippingMethodChange(
-            componentId: componentId,
-            shippingMethod: shippingMethod.toDTO(currencyCode: payment.currencyCode),
-            currentSummaryItems: payment.summaryItems.map { $0.toDTO(currencyCode: payment.currencyCode) },
-            completion: { result in
-                switch result {
-                case let .success(update):
-                    completion(update.toPKPaymentRequestShippingMethodUpdate())
-                case .failure:
-                    completion(PKPaymentRequestShippingMethodUpdate(paymentSummaryItems: payment.summaryItems))
-                }
-            }
-        )
-    }
-
-    @available(iOS 15.0, *)
-    func didUpdate(
-        couponCode: String,
-        for payment: ApplePayPayment,
-        completion: @escaping (PKPaymentRequestCouponCodeUpdate) -> Void
-    ) {
-        completion(PKPaymentRequestCouponCodeUpdate(paymentSummaryItems: payment.summaryItems))
     }
 }
